test(module): cannot add two lectures with same name
The past test ensured that we cannot add the same lecture (i.e. with same
attributes) twice in the same module. This test is different, it focuses
only in the name of the lecture. Two lectures can have same name and still
be treated as different because of the other attributes, so this test
comes in.