feat(module): cannot add two lectures with same name
The "add" method of the module class don't look if there is some identical
lecture in the array anymore, it looks only if there is some lecture with
same description.

If two lectures are equal, they have the same description, but it is
possible to have different lectures with same description. Therefore
the analysis of the description is enough, because it covers the cases
in which the lectures are equal.