test(module): rearrange lecture order
The test uses the unimplemented methods "move" and "position". The first
will move the position of the lecture in the array of lectures inside the
module and the second will receive a lecture and will return its position
inside the array of lectures.

The two methods receive a lecture and must search for another lecture
inside the array that equals the first lecture. The positions start at
one, unlike the real indexes of the array.