test(course): rearrange module order
The test uses the unimplemented methods "move" and "position". The first
will move the position of the module in the array of modules inside the
course and the second will receive a module and will return its position
inside the array of modules.

The two methods receive a module and must search for another module
inside the array that equals the first module. The positions start at
one, unlike the real indexes of the array.