test(lecture): further material
A lecture should contain a link to the main video, but it may also provide
further contents if it is the will of the teatcher. Then, the lecture
class must implement an interface to your interior array of materials.

The Material interface must be created, and a lecture should be able
to add, remove and see if contains objects from classes that implements
the Material interface. The test also uses a class that must implement
the interface, the Pdf.